Moved contributions to contributor.ts config file
